Added sninfo/sfinfo; Fixed unknown data source
Added functions to give output as to what stations/data are in surface
and sounding files. It requires a separate loop through the data to test
for null pointers, but the cost is minimal.

Added better handling of when an unknown data source is loaded. Previous
behavior was to throw and exception and quit. Now the source is set to
Unknown and the user is given a warning.

Improve some method docstrings to be more clear about expected input and
what default parameters will do.